+++ src/main/java/com/supermartijn642/formations/tools/loot_table/ContainerOpenIntercept.java
@@ -36,7 +36,7 @@ public class ContainerOpenIntercept {
 
             BlockEntity entity = world.getBlockEntity(event.getPos());
             if(entity instanceof RandomizableContainerBlockEntity && (!player.isShiftKeyDown() || ((RandomizableContainerBlockEntity)entity).lootTable != null)){
-                FormationsDev.CHANNEL.sendToPlayer(player, new OpenLootTableScreenPacket(entity.getBlockPos(), ((RandomizableContainerBlockEntity)entity).lootTable));
+                FormationsDev.CHANNEL.sendToPlayer(player, new OpenLootTableScreenPacket(entity.getBlockPos(), ((RandomizableContainerBlockEntity)entity).lootTable.location()));
                 event.setCanceled(true);
                 event.setCancellationResult(InteractionResult.FAIL);
             }